:art: Refactor project structure by splitting in packages.
