+++ CONTRIBUTING.md
@@ -6,10 +6,18 @@ Welcome to Imerologio CLI project !
 Imerologio CLI is a Go application so make sure you have a development environment with Go 1.2+.  
 [See the install instructions for Go](http://golang.org/doc/install.html).
 
+The project is structured like adviced in [golang-standards](https://github.com/golang-standards/project-layout).
+
 ### Dependencies
 Install the following dependencies with `go get XXX`:
 - [ishell](https://github.com/abiosoft/ishell) to ensure an interactive CLI : `go get github.com/fatih/color`
 - [Color](https://github.com/fatih/color) in shell : `go get github.com/fatih/color`
+
+### Build
+Launch `go build -o imerologio-cli cmd/imerologio_cli/imerologio_cli.go` to generate an executable named `imerologio-cli` at the root.
+
+### Launch app
+Simply run `./imerologio-cli` once build finished.
 
 ## About releasing
 This product is built using [goreleaser](https://goreleaser.com/), take a look at the config at [.goreleaser.yml](.goreleaser.yml).
