+++ pkg/ime_types/answers_type.go
@@ -1,0 +1,6 @@
+package ime_types
+
+type Answers struct {
+	Name string
+	Path string
+}
